requirements.txt
# Inference
torch
transformers

fastapi
pydantic
psycopg2
uvicorn
python-dotenv
SQLAlchemy
pydantic[email]
passlib[bcrypt]
PyJWT

vllm